library/requirements.txt
django==5.2.14
djangorestframework==3.17.1
Pillow==12.2.0
requests==2.33.1
django-cors-headers==4.9.0
